refactor: rename src/config to src/default
- config will be used for new Configuration classes
