+++ src/brisk/__init__.py
@@ -1,7 +1,7 @@
-from brisk.config.RegressionAlgorithms import REGRESSION_ALGORITHMS
-from brisk.config.RegressionMetrics import REGRESSION_METRICS
-from brisk.config.ClassificationAlgorithms import CLASSIFICATION_ALGORITHMS
-from brisk.config.ClassificationMetrics import CLASSIFICATION_METRICS 
+from brisk.defaults.RegressionAlgorithms import REGRESSION_ALGORITHMS
+from brisk.defaults.RegressionMetrics import REGRESSION_METRICS
+from brisk.defaults.ClassificationAlgorithms import CLASSIFICATION_ALGORITHMS
+from brisk.defaults.ClassificationMetrics import CLASSIFICATION_METRICS 
 
 from brisk.data.DataManager import DataManager
 from brisk.evaluation.MetricManager import MetricManager
