+++ src/brisk/defaults/ClassificationAlgorithms.py
