+++ src/brisk/defaults/RegressionAlgorithms.py
